move app logic to factory view
